Added storage and VAC checks for suspects
After analyzing the demo csgo-overwatcher will present a list of all
players where the user has to choose which one is the victim (or, if the
suspect is innocent). Some data about the suspect will be stored in a
local file called suspects.json with the current VAC status.

All new suspects will be appended to the file. Calling csgo-overwatcher
with the suspects.json as first argument (instead of a .dem file) will
check the VAC status of all stored suspects that are not listed as
already banned. If a user has been banned, the status will be updated in
the suspects.json file.

This allows to easily track if a suspect has been banned and, depending
on how frequently it will run, roughly when the ban happened.